requirements.txt
flask
mariadb
mysql-connector-python
ollama
chromadb
mysql-connector-python>=8.0.0
